pyproject.toml
[tool.poetry]
name = "kwurify"
version = "0.1.0"
description = "summer project intended to create playlists for radio DJs at KWUR"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spotipy = "^2.23.0"
mkdocs = "^1.4.3"
python-dotenv = "^1.0.0"
flask = "^2.3.2"
pandas = "^2.0.1"
ipykernel = "^6.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
